requirements.txt
streamlit
pandas
PyPDF2
python-docx
xlsxwriter
pytesseract
pillow
pdfplumber
pymupd
requests
openpyxl
scikit-learn
matplotlib
plotly
tqdm
google-generativeai